qwqw

+++ src/main/java/com/example/diploma/dto/AdDto.java
@@ -7,15 +7,20 @@ import lombok.Data;
 @Schema(description = "Краткая информация об объявлении")
 public class AdDto {
     @Schema(description = "ID объявления")
-    private Integer pk;
+    private Long pk; // Было Integer
+
     @Schema(description = "ID автора")
-    private Integer author; // <-- Изменено с authorId на author
+    private Integer author;
+
     @Schema(description = "URL изображения")
     private String image;
+
     @Schema(description = "Цена", example = "10000")
     private Integer price;
+
     @Schema(description = "Заголовок", example = "Продам велосипед")
     private String title;
+
     @Schema(description = "Описание", example = "Описание товара")
     private String description;
 }
